Lots of stuff
Reversed the logic, now when clicking on a villager he gives you the
prices for more of that item (ie: click on him with a diamond and he'll
tell you how many coal he'd take to give you a diamond).
Added variable prices for tools and weapons, the more damaged a tool or
weapon is, the less it's worth.